refac: ajuste na home page + adicao da funcionalidade de fauna local
